refactor(model): split the repository file by concern
model.go held the Model repository's CRUD alongside the retirement
bookkeeping and the prune of retired rows. Same package, pure moves
with their doc comments:

- retire.go: DisabledModelRef, MissingScanThreshold,
  RecordMissingModels, SetEnabled, AutoRetireIfConfirmed,
  RevertAutoRetire
- prune.go: PrunedModel, PruneRetired, notFlappedSince,
  selectPruneCandidates, deletePruneCandidates, aliveModelIDs

model.go keeps the types, the constructor, Upsert, the reads, the
deletes and Update (983 -> 563 lines) and leaves the size allowlist.
The tests were already split the same way (model_crud_test,
model_prune_test, scan_test) and are unchanged.
